Rename crypto->ikecrypto to prevent naming conflicts on OSX (different filename case is ignored)
